+++ src/shared/icons/Check.tsx
