+++ src/shared/icons/Copy.tsx
