+++ src/shared/icons/EqualizeH.tsx
