+++ src/shared/icons/EqualizeL.tsx
