+++ src/shared/icons/Minimize.tsx
