+++ src/shared/icons/More.tsx
