+++ src/shared/icons/Settings.tsx
