+++ src/shared/icons/Sun.tsx
