+++ src/shared/icons/Trash.tsx
