+++ src/shared/presets/stripe.json
