+++ src/shared/presets/stripe_dashboard.json
