+++ src/store/palette/palette.test.ts
@@ -1,4 +1,4 @@
-import { HexPalette } from '../../shared/types'
+import { HexPalette, spaceName } from '../../shared/types'
 import { exportToHexPalette, parseHexPalette } from './converters'
 
 const hexPalette: HexPalette = {
@@ -11,7 +11,7 @@ const hexPalette: HexPalette = {
 }
 
 test('Parsing hex palette', () => {
-  const palette = parseHexPalette(hexPalette, 'cielch')
+  const palette = parseHexPalette(hexPalette, spaceName.cielch)
   const hex = exportToHexPalette(palette)
   expect(hex).toEqual(hexPalette)
   expect(palette.colors.length).toEqual(hexPalette.hues.length)
